optimize: Transfer the operation of adding configuration class from 'AutoConfiguration' to 'EnvironmentPostProcessor'
