+++ seata-spring-autoconfigure/seata-spring-autoconfigure-client/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,5 @@
 # Auto Configure
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  io.seata.spring.boot.autoconfigure.SeataClientPropertiesAutoConfiguration,\
-  io.seata.spring.boot.autoconfigure.SeataTCCFenceAutoConfiguration
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=io.seata.spring.boot.autoconfigure.SeataTCCFenceAutoConfiguration
+
+# Environment Post Processors
+org.springframework.boot.env.EnvironmentPostProcessor=io.seata.spring.boot.autoconfigure.SeataClientEnvironmentPostProcessor
