+++ seata-spring-autoconfigure/seata-spring-autoconfigure-core/src/main/java/io/seata/spring/boot/autoconfigure/SeataCoreAutoConfiguration.java
@@ -1,0 +1,41 @@
+/*
+ *  Copyright 1999-2019 Seata.io Group.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package io.seata.spring.boot.autoconfigure;
+
+import io.seata.spring.boot.autoconfigure.provider.SpringApplicationContextProvider;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+import static io.seata.common.Constants.BEAN_NAME_SPRING_APPLICATION_CONTEXT_PROVIDER;
+import static io.seata.spring.boot.autoconfigure.StarterConstants.SEATA_PREFIX;
+
+/**
+ * @author xingfudeshi@gmail.com
+ */
+@ConditionalOnProperty(prefix = SEATA_PREFIX, name = "enabled", havingValue = "true", matchIfMissing = true)
+@ComponentScan(basePackages = "io.seata.spring.boot.autoconfigure.properties")
+@Configuration
+public class SeataCoreAutoConfiguration {
+
+    @Bean(BEAN_NAME_SPRING_APPLICATION_CONTEXT_PROVIDER)
+    @ConditionalOnMissingBean(name = {BEAN_NAME_SPRING_APPLICATION_CONTEXT_PROVIDER})
+    public SpringApplicationContextProvider springApplicationContextProvider() {
+        return new SpringApplicationContextProvider();
+    }
+}
