+++ seata-spring-autoconfigure/seata-spring-autoconfigure-server/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
-# Auto Configure
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=io.seata.spring.boot.autoconfigure.SeataServerPropertiesAutoConfiguration
+# Environment Post Processors
+org.springframework.boot.env.EnvironmentPostProcessor=io.seata.spring.boot.autoconfigure.SeataServerEnvironmentPostProcessor
